googlecast: Replaced loadedmedata event with qualityrequested
When a new quality is requested, plyr changes quality and continues
the video from where it left off. Hooking in to loadedmetadata will
instead trigger when the new video's metadata is ready and as a
result, the video state is lost.

By using qualityrequested event, we are able to issue the loadMedia
request with the right properties and ensure that the remote player
behaves the same way plyr does.